chore: add compiler flags to the project and address the issues

+++ src/common/program.cpp
@@ -20,7 +20,7 @@ void Program::linkProgram() const
     {
         int logLength = 0;
         glGetProgramiv(programId_, GL_INFO_LOG_LENGTH, &logLength);
-        char *message = static_cast<char *>(alloca(logLength * sizeof(char)));
+        char *message = static_cast<char *>(alloca(static_cast<long unsigned int>(logLength) * sizeof(char)));
         glGetProgramInfoLog(programId_, logLength, &logLength, message);
         throw std::runtime_error(message);
     }
